+++ src/test/clojure/clojure/core/typed/test/async_walkthrough.clj
@@ -2,7 +2,8 @@
 
 ;; The clojure.core.async namespace contains the public API.
 (ns clojure.core.typed.test.async-walkthrough
-  (:require [clojure.core.typed :as t :refer [inst check-ns doseq> dotimes>]]
+  (:require [clojure.core.typed :as t :refer [inst check-ns doseq> dotimes> ann-form
+                                              Seqable]]
             [clojure.core.typed.async :refer [chan> go> Chan sliding-buffer> dropping-buffer>]]
             [clojure.core.async :as async :refer :all]))