+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 0.1.19
 - Start supporting core.async
-  - see c.c.t.test.async-* for examples
-  - c.c.t.async is the companion namespace
+  - see c.c.t.test.async-{go,alts} for examples
+  - c.c.t.async is the companion namespace. Just require it
+    and the annotations will be picked up in your core.async ns
 - Deprecate implicit optional keyword parameters
   - old syntax: [args & {opts} :mandatory {mand...} -> t]
   - new syntax (interchangable keyword args): [& :optional {opts} :mandatory {mand...} -> t]
@@ -23,6 +24,8 @@
   - (EmptyCount NonEmptyCount)
     - (ExactCount 0) and (CountRange 1)
 - Update base env with new aliases
+- Fixes for Extends type
+  - subtyping, cs gen
 
 0.1.18
 - Update to tools.namespace SNAPSHOT, avoids StackOverflow error